requirements.txt
Flask-Cors
Flask
pandas
scikit-learn
openai
requests
u8darts[all]

